Changed implementation of the Chords tab
I was able to remove a for loop by changing how the program finds the note for a given chord. I also moved the switch statement from mainwinow.cpp over to chords.cpp to clean up mainwindow.cpp and allow the code to be reused for a chord progression generator for later.